Fix handling of chrome
