+++ lib/commands/context.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import logger from '../logger';
+import log from '../logger';
 import Chromedriver from 'appium-chromedriver';
 import PortFinder from 'portfinder';
 import B from 'bluebird';
@@ -32,7 +32,7 @@ commands.getContexts = async function () {
       this.opts.androidDeviceSocket);
   }
   this.contexts = _.union([NATIVE_WIN], webviews);
-  logger.debug(`Available contexts: ${JSON.stringify(this.contexts)}`);
+  log.debug(`Available contexts: ${JSON.stringify(this.contexts)}`);
   return this.contexts;
 };
 
@@ -69,7 +69,7 @@ helpers.switchContext = async function (name) {
     // to true then kill chromedriver session using stopChromedriverProxies or
     // else simply suspend proxying to the latter
     if (this.opts.recreateChromeDriverSessions) {
-      logger.debug("recreateChromeDriverSessions set to true; killing existing chromedrivers");
+      log.debug("recreateChromeDriverSessions set to true; killing existing chromedrivers");
       this.stopChromedriverProxies();
     } else {
       this.suspendChromedriverProxy();
@@ -101,13 +101,13 @@ helpers.isWebContext = function () {
 
 // Turn on proxying to an existing Chromedriver session or a new one
 helpers.startChromedriverProxy = async function (context) {
-  logger.debug(`Connecting to chrome-backed webview context '${context}'`);
+  log.debug(`Connecting to chrome-backed webview context '${context}'`);
 
   let cd;
   if (this.sessionChromedrivers[context]) {
     // in the case where we've already set up a chromedriver for a context,
     // we want to reconnect to it, not create a whole new one
-    logger.debug(`Found existing Chromedriver for context '${context}'. Using it.`);
+    log.debug(`Found existing Chromedriver for context '${context}'. Using it.`);
     cd = this.sessionChromedrivers[context];
     await setupExistingChromedriver(cd);
   } else {
@@ -119,8 +119,8 @@ helpers.startChromedriverProxy = async function (context) {
         opts.chromeAndroidPackage = androidPackage[1];
       }
     }
-    cd = await setupNewChromedriver(opts, this.adb.curDeviceId,
-                                    this.adb);
+
+    cd = await this.setupNewChromedriver(opts, this.adb.curDeviceId, this.adb);
     // bind our stop/exit handler, passing in context so we know which
     // one stopped unexpectedly
     cd.on(Chromedriver.EVENT_CHANGED, (msg) => {
@@ -146,7 +146,7 @@ helpers.suspendChromedriverProxy = function () {
 
 // Handle an out-of-band Chromedriver stop event
 helpers.onChromedriverStop = async function (context) {
-  logger.warn(`Chromedriver for context ${context} stopped unexpectedly`);
+  log.warn(`Chromedriver for context ${context} stopped unexpectedly`);
   if (context === this.curContext) {
     // we exited unexpectedly while automating the current context and so want
     // to shut down the session and respond with an error
@@ -155,7 +155,7 @@ helpers.onChromedriverStop = async function (context) {
   } else {
     // if a Chromedriver in the non-active context barfs, we don't really
     // care, we'll just make a new one next time we need the context.
-    logger.warn("Chromedriver quit unexpectedly, but it wasn't the active " +
+    log.warn("Chromedriver quit unexpectedly, but it wasn't the active " +
                 "context, ignoring");
     delete this.sessionChromedrivers[context];
   }
@@ -167,13 +167,13 @@ helpers.stopChromedriverProxies = async function () {
   this.suspendChromedriverProxy(); // make sure we turn off the proxy flag
   for (let context of _.keys(this.sessionChromedrivers)) {
     let cd = this.sessionChromedrivers[context];
-    logger.debug(`Stopping chromedriver for context ${context}`);
+    log.debug(`Stopping chromedriver for context ${context}`);
     // stop listening for the stopped state event
     cd.removeAllListeners(Chromedriver.EVENT_CHANGED);
     try {
       await cd.stop();
     } catch (err) {
-      logger.warn(`Error stopping Chromedriver: ${err.message}`);
+      log.warn(`Error stopping Chromedriver: ${err.message}`);
     }
     delete this.sessionChromedrivers[context];
   }
@@ -181,6 +181,66 @@ helpers.stopChromedriverProxies = async function () {
 
 helpers.isChromedriverContext = function (viewName) {
   return _.includes(viewName, WEBVIEW_WIN) || viewName === CHROMIUM_WIN;
+};
+
+helpers.shouldDismissChromeWelcome = function shouldDismissChromeWelcome () {
+  return !_.isUndefined(this.opts.chromeOptions) &&
+    _.isArray(this.opts.chromeOptions.args) &&
+    this.opts.chromeOptions.args.indexOf('--no-first-run') !== -1;
+};
+
+helpers.dismissChromeWelcome = async function dismissChromeWelcome () {
+  log.info("Trying to dismiss Chrome welcome");
+  let activity = await this.getCurrentActivity();
+  if (activity !== "org.chromium.chrome.browser.firstrun.FirstRunActivity") {
+    log.info("Chrome welcome dialog never showed up! Continuing");
+    return;
+  }
+  let el = await this.findElOrEls('id', 'com.android.chrome:id/terms_accept', false);
+  await this.click(el.ELEMENT);
+  try {
+    let el = await this.findElOrEls('id', 'com.android.chrome:id/negative_button', false);
+    await this.click(el.ELEMENT);
+  } catch (e) {
+    // DO NOTHING, THIS DEVICE DIDNT LAUNCH THE SIGNIN DIALOG
+    // IT MUST BE A NON GMS DEVICE
+    log.warn(`This device didnt show Chrome SignIn dialog, ${e.message}`);
+  }
+};
+
+helpers.startChromeSession = async function startChromeSession () {
+  log.info("Starting a chrome-based browser session");
+  let opts = _.cloneDeep(this.opts);
+  opts.chromeUseRunningApp = false;
+
+  const knownPackages = ["org.chromium.chrome.shell",
+    "com.android.chrome",
+    "com.chrome.beta",
+    "org.chromium.chrome",
+    "org.chromium.webview_shell"];
+
+  if (!_.includes(knownPackages, this.opts.appPackage)) {
+    opts.chromeAndroidActivity = this.opts.appActivity;
+  }
+  this.chromedriver = await this.setupNewChromedriver(opts, this.adb.curDeviceId, this.adb);
+  this.chromedriver.on(Chromedriver.EVENT_CHANGED, (msg) => {
+    if (msg.state === Chromedriver.STATE_STOPPED) {
+      this.onChromedriverStop(CHROMIUM_WIN);
+    }
+  });
+
+  // Now that we have a Chrome session, we ensure that the context is
+  // appropriately set and that this chromedriver is added to the list
+  // of session chromedrivers so we can switch back and forth
+  this.curContext = CHROMIUM_WIN;
+  this.sessionChromedrivers[CHROMIUM_WIN] = this.chromedriver;
+  this.proxyReqRes = this.chromedriver.proxyReq.bind(this.chromedriver);
+  this.jwpProxyActive = true;
+
+  if (this.shouldDismissChromeWelcome()) {
+    // dismiss Chrome welcome dialog
+    await this.dismissChromeWelcome();
+  }
 };
 
 
@@ -192,19 +252,19 @@ async function setupExistingChromedriver (chromedriver) {
   // check the status by sending a simple window-based command to ChromeDriver
   // if there is an error, we want to recreate the ChromeDriver session
   if (!await chromedriver.hasWorkingWebview()) {
-    logger.debug("ChromeDriver is not associated with a window. " +
+    log.debug("ChromeDriver is not associated with a window. " +
                  "Re-initializing the session.");
     await chromedriver.restart();
   }
   return chromedriver;
 }
 
-async function setupNewChromedriver (opts, curDeviceId, adb) {
+helpers.setupNewChromedriver = async function setupNewChromedriver (opts, curDeviceId, adb) {
   // if a port wasn't given, pick a random available one
   if (!opts.chromeDriverPort) {
     let getPort = B.promisify(PortFinder.getPort, {context: PortFinder});
     opts.chromeDriverPort = await getPort();
-    logger.debug(`A port was not given, using random port: ${opts.chromeDriverPort}`);
+    log.debug(`A port was not given, using random port: ${opts.chromeDriverPort}`);
   }
   let chromeArgs = {
     port: opts.chromeDriverPort,
@@ -220,7 +280,7 @@ async function setupNewChromedriver (opts, curDeviceId, adb) {
   // and strip the prefix
   for (let opt of _.keys(opts)) {
     if (opt.endsWith(':chromeOptions')) {
-      logger.warn(`Merging '${opt}' into 'chromeOptions'. This may cause unexpected behavior`);
+      log.warn(`Merging '${opt}' into 'chromeOptions'. This may cause unexpected behavior`);
       _.merge(opts.chromeOptions, opts[opt]);
     }
   }
@@ -255,7 +315,8 @@ async function setupNewChromedriver (opts, curDeviceId, adb) {
   caps = webviewHelpers.decorateChromeOptions(caps, opts, curDeviceId);
   await chromedriver.start(caps);
   return chromedriver;
-}
+};
+const setupNewChromedriver = helpers.setupNewChromedriver;
 
 Object.assign(extensions, commands, helpers);
 export { commands, helpers, setupNewChromedriver };
