+++ lib/commands/index.js
@@ -14,6 +14,7 @@ import executeCmds from "./execute";
 import shellCmds from "./shell";
 import appManagementCmds from './app-management';
 
+
 let commands = {};
 Object.assign(
   commands,
