+++ lib/driver.js
@@ -1,10 +1,7 @@
 import { BaseDriver, DeviceSettings } from 'appium-base-driver';
-import Chromedriver from 'appium-chromedriver';
 import desiredConstraints from './desired-caps';
 import commands from './commands/index';
-import * as contextCommands from './commands/context';
 import helpers from './android-helpers';
-import { CHROMIUM_WIN } from './webview-helpers';
 import log from './logger';
 import _ from 'lodash';
 import { DEFAULT_ADB_PORT } from 'appium-adb';
@@ -276,31 +273,6 @@ class AndroidDriver extends BaseDriver {
     await this.initAutoWebview();
   }
 
-  shouldDismissChromeWelcome () {
-    return !_.isUndefined(this.opts.chromeOptions) &&
-      _.isArray(this.opts.chromeOptions.args) &&
-      this.opts.chromeOptions.args.indexOf('--no-first-run') !== -1;
-  }
-
-  async dismissChromeWelcome () {
-    log.info("Trying to dismiss Chrome welcome");
-    let activity = await this.getCurrentActivity();
-    if (activity !== "org.chromium.chrome.browser.firstrun.FirstRunActivity") {
-      log.info("Chrome welcome dialog never showed up! Continuing");
-      return;
-    }
-    let el = await this.findElOrEls('id', 'com.android.chrome:id/terms_accept', false);
-    await this.click(el.ELEMENT);
-    try {
-      let el = await this.findElOrEls('id', 'com.android.chrome:id/negative_button', false);
-      await this.click(el.ELEMENT);
-    } catch (e) {
-      // DO NOTHING, THIS DEVICE DIDNT LAUNCH THE SIGNIN DIALOG
-      // IT MUST BE A NON GMS DEVICE
-      log.warn(`This device didnt show Chrome SignIn dialog, ${e.message}`);
-    }
-  }
-
   async initAutoWebview () {
     if (this.opts.autoWebview) {
       let viewName = this.defaultWebviewName();
@@ -344,42 +316,6 @@ class AndroidDriver extends BaseDriver {
     // install to fail. And before running the app.
     if (!_.isUndefined(this.opts.sharedPreferences)) {
       await this.setSharedPreferences(this.opts);
-    }
-  }
-
-  async startChromeSession () {
-    log.info("Starting a chrome-based browser session");
-    let opts = _.cloneDeep(this.opts);
-    opts.chromeUseRunningApp = false;
-
-    const knownPackages = ["org.chromium.chrome.shell",
-                           "com.android.chrome",
-                           "com.chrome.beta",
-                           "org.chromium.chrome",
-                           "org.chromium.webview_shell"];
-
-    if (!_.includes(knownPackages, this.opts.appPackage)) {
-      opts.chromeAndroidActivity = this.opts.appActivity;
-    }
-    this.chromedriver = await contextCommands.setupNewChromedriver(opts, this.adb.curDeviceId,
-                                                                   this.adb);
-    this.chromedriver.on(Chromedriver.EVENT_CHANGED, (msg) => {
-      if (msg.state === Chromedriver.STATE_STOPPED) {
-        this.onChromedriverStop(CHROMIUM_WIN);
-      }
-    });
-
-    // Now that we have a Chrome session, we ensure that the context is
-    // appropriately set and that this chromedriver is added to the list
-    // of session chromedrivers so we can switch back and forth
-    this.curContext = CHROMIUM_WIN;
-    this.sessionChromedrivers[CHROMIUM_WIN] = this.chromedriver;
-    this.proxyReqRes = this.chromedriver.proxyReq.bind(this.chromedriver);
-    this.jwpProxyActive = true;
-
-    if (this.shouldDismissChromeWelcome()) {
-      // dismiss Chrome welcome dialog
-      await this.dismissChromeWelcome();
     }
   }
